+++ Warehouse/Warehouse.conf.in
@@ -2,3 +2,11 @@ precondition = ["Platform"]
 callsign = "org.rdk.Warehouse"
 autostart = "false"
 startuporder = "@PLUGIN_WAREHOUSE_STARTUPORDER@"
+
+configuration = JSON()
+rootobject = JSON()
+
+rootobject.add("mode", "@PLUGIN_WAREHOUSE_MODE@")
+rootobject.add("locator", "lib@PLUGIN_IMPLEMENTATION@.so")
+
+configuration.add("root", rootobject)
