+++ Warehouse/Warehouse.config
@@ -5,3 +5,12 @@ set (callsign "org.rdk.Warehouse")
 if(PLUGIN_WAREHOUSE_STARTUPORDER)
 set (startuporder ${PLUGIN_WAREHOUSE_STARTUPORDER})
 endif()
+
+map()
+    key(root)
+    map()
+        kv(mode ${PLUGIN_WAREHOUSE_MODE})
+        kv(locator lib${PLUGIN_IMPLEMENTATION}.so)
+    end()
+end()
+ans(configuration)
