+++ Warehouse/Warehouse.h
@@ -1,148 +1,108 @@
 /**
-* If not stated otherwise in this file or this component's LICENSE
-* file the following copyright and licenses apply:
-*
-* Copyright 2019 RDK Management
-*
-* Licensed under the Apache License, Version 2.0 (the "License");
-* you may not use this file except in compliance with the License.
-* You may obtain a copy of the License at
-*
-* http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-* See the License for the specific language governing permissions and
-* limitations under the License.
-**/
+ * If not stated otherwise in this file or this component's LICENSE
+ * file the following copyright and licenses apply:
+ *
+ * Copyright 2025 RDK Management
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ **/
 
 #pragma once
 
-#include <thread>
-
 #include "Module.h"
-#include "UtilsThreadRAII.h"
-#include "libIARM.h"
+#include <interfaces/IWarehouse.h>
+#include <interfaces/json/JWarehouse.h>
+#include <interfaces/json/JsonData_Warehouse.h>
+#include "UtilsLogging.h"
+#include "tracing/Logging.h"
 
-namespace WPEFramework {
-
-    namespace Plugin {
-
-#ifdef HAS_FRONT_PANEL
-        class Warehouse;
-        class LedInfo
+namespace WPEFramework 
+{
+    namespace Plugin
+    {
+        class Warehouse : public PluginHost::IPlugin, public PluginHost::JSONRPC 
         {
-        private:
-            LedInfo() = delete;
-            LedInfo& operator=(const LedInfo& RHS) = delete;
+            private:
+                class Notification : public RPC::IRemoteConnection::INotification, public Exchange::IWarehouse::INotification
+                {
+                    private:
+                        Notification() = delete;
+                        Notification(const Notification&) = delete;
+                        Notification& operator=(const Notification&) = delete;
 
-        public:
-            LedInfo(Warehouse* wh)
-            : m_warehouse(wh)
-            {
-            }
-            LedInfo(const LedInfo& copy)
-            : m_warehouse(copy.m_warehouse)
-            {
-            }
-            ~LedInfo() {}
+                    public:
+                    explicit Notification(Warehouse* parent) 
+                        : _parent(*parent)
+                        {
+                            ASSERT(parent != nullptr);
+                        }
 
-            inline bool operator==(const LedInfo& RHS) const
-            {
-                return(m_warehouse == RHS.m_warehouse);
-            }
+                        virtual ~Notification()
+                        {
+                        }
 
-        public:
-            uint64_t Timed(const uint64_t scheduledTime);
+                        BEGIN_INTERFACE_MAP(Notification)
+                        INTERFACE_ENTRY(Exchange::IWarehouse::INotification)
+                        INTERFACE_ENTRY(RPC::IRemoteConnection::INotification)
+                        END_INTERFACE_MAP
 
-        private:
-            Warehouse* m_warehouse;
-        };
-#endif
+                        void Activated(RPC::IRemoteConnection*) override
+                        {
+                           
+                        }
 
-        // This is a server for a JSONRPC communication channel. 
-        // For a plugin to be capable to handle JSONRPC, inherit from PluginHost::JSONRPC.
-        // By inheriting from this class, the plugin realizes the interface PluginHost::IDispatcher.
-        // This realization of this interface implements, by default, the following methods on this plugin
-        // - exists
-        // - register
-        // - unregister
-        // Any other methood to be handled by this plugin  can be added can be added by using the
-        // templated methods Register on the PluginHost::JSONRPC class.
-        // As the registration/unregistration of notifications is realized by the class PluginHost::JSONRPC,
-        // this class exposes a public method called, Notify(), using this methods, all subscribed clients
-        // will receive a JSONRPC message as a notification, in case this method is called.
-        class Warehouse : public PluginHost::IPlugin, public PluginHost::JSONRPC {
-        private:
+                        void Deactivated(RPC::IRemoteConnection *connection) override
+                        {
+                            _parent.Deactivated(connection);
+                        }
 
-            // We do not allow this plugin to be copied !!
-            Warehouse(const Warehouse&) = delete;
-            Warehouse& operator=(const Warehouse&) = delete;
+                        void ResetDone(const bool success, const string& error) override
+                        {
+                            LOGINFO("ResetDone");
+                            Exchange::JWarehouse::Event::ResetDone(_parent, success,error);
+                        }
 
-            void resetDevice(bool suppressReboot, const string& resetType = string());
-            void internalReset(JsonObject& response);
-            void lightReset(JsonObject& response);
-            void isClean(int age, JsonObject& response);
-            bool executeHardwareTest() const;
-            bool getHardwareTestResults(string& testResults) const;
+                    private:
+                        Warehouse& _parent;
+                };
 
-            //Begin methods
-            uint32_t resetDeviceWrapper(const JsonObject& parameters, JsonObject& response);
-            uint32_t internalResetWrapper(const JsonObject& parameters, JsonObject& response);
-            uint32_t lightResetWrapper(const JsonObject& parameters, JsonObject& response);
-            uint32_t isCleanWrapper(const JsonObject& parameters, JsonObject& response);
-            uint32_t executeHardwareTestWrapper(const JsonObject& parameters, JsonObject& response);
-            uint32_t getHardwareTestResultsWrapper(const JsonObject& parameters, JsonObject& response);
-            //End methods
+                public:
+                    Warehouse(const Warehouse&) = delete;
+                    Warehouse& operator=(const Warehouse&) = delete;
 
-	    static void dsWareHouseOpnStatusChanged(const char *owner, IARM_EventId_t eventId, void *data, size_t len);
-        public:
-            Warehouse();
-            virtual ~Warehouse();
-            //IPlugin methods
-            virtual const string Initialize(PluginHost::IShell* service) override;
-            virtual void Deinitialize(PluginHost::IShell* service) override;
-            virtual string Information() const override { return {}; }
+                    Warehouse();
+                    virtual ~Warehouse();
 
-            void onSetFrontPanelStateTimer();
+                    BEGIN_INTERFACE_MAP(Warehouse)
+                    INTERFACE_ENTRY(PluginHost::IPlugin)
+                    INTERFACE_ENTRY(PluginHost::IDispatcher)
+                    INTERFACE_AGGREGATE(Exchange::IWarehouse, _warehouse)
+                    END_INTERFACE_MAP
 
-            BEGIN_INTERFACE_MAP(Warehouse)
-            INTERFACE_ENTRY(PluginHost::IPlugin)
-            INTERFACE_ENTRY(PluginHost::IDispatcher)
-            END_INTERFACE_MAP
+                    //  IPlugin methods
+                    // -------------------------------------------------------------------------------------------------------
+                    const string Initialize(PluginHost::IShell* service) override;
+                    void Deinitialize(PluginHost::IShell* service) override;
+                    string Information() const override;
 
-        private:
-            /*returns the UTC date and time in format DD.MM.YYYY_HH.MM.SS format eg : 11.08.2023_18:47:47*/
-            void getDateAndTime(string& utcDateTime);
-            /*gets the SD card mount path by reading /proc/mounts, returns true on success, false otherwise*/
-            bool getSDCardMountPath(string&);
-            /*Adds 0 to file /opt/.rebootFlag*/
-            void resetWarehouseRebootFlag();
-            void InitializeIARM();
-            void DeinitializeIARM();
+                private:
+                    void Deactivated(RPC::IRemoteConnection* connection);
 
-
-            Utils::ThreadRAII m_resetThread;
-
-#ifdef HAS_FRONT_PANEL
-            Core::TimerType<LedInfo> m_ledTimer;
-            LedInfo m_ledInfo;
-            int m_ledTimerIteration;
-            int m_ledState;
-#endif
-        public:
-            static Warehouse* _instance;
-
-            uint32_t processColdFactoryReset();
-            uint32_t processFactoryReset();
-            uint32_t processWareHouseReset();
-            uint32_t processWHReset();
-            uint32_t processWHResetNoReboot();
-            uint32_t processWHClear();
-            uint32_t processWHClearNoReboot();
-            uint32_t processUserFactoryReset();
-
-        };
-	} // namespace Plugin
+                private:
+                    PluginHost::IShell* _service{};
+                    uint32_t _connectionId{};
+                    Exchange::IWarehouse* _warehouse{};
+                    Core::Sink<Notification> _warehouseNotification;
+       };
+    } // namespace Plugin
 } // namespace WPEFramework
